renamed tensorflow1 scripts to tensorflow
